requirements.txt
pytorch==1.12.0
scikit-learn==1.0.2
numpy==1.22.3
matplotlib==3.3.4
scipy==1.7.3
pandas==1.4.2
psutil==5.8.0 